+++ packages/web/src/components/ProjectCard/ProjectCard.tsx
@@ -1,0 +1,20 @@
+import { Heading, Text, Flex } from '@chakra-ui/react';
+import { Project } from '@hangar/shared';
+
+type ProjectCardProps = {
+  project: Project;
+};
+
+export const ProjectCard: React.FC<ProjectCardProps> = ({ project }) => {
+  const { name, description, location, repoUrl } = project;
+
+  return (
+    <Flex>
+      <Heading size="md">{name}</Heading>
+
+      <Text>{description}</Text>
+      <Text>{location}</Text>
+      <Text>{repoUrl}</Text>
+    </Flex>
+  );
+};
