+++ packages/web/src/components/ProjectCard/index.tsx
@@ -1,0 +1,1 @@
+export * from './ProjectCard';
