--- packages/web/src/components/ProjectList/index.tsx
@@ -1,1 +1,0 @@
-export * from './ProjectList';
